refactor: renomear módulo csv-json para serialization
- Pasta e arquivos renomeados: csv-json.* → serialization.*
- Componentes: CsvJson* → Serialization*, selectors atualizados
- Rota da URL: /conversores/csv-json → /conversores/serialization
- ID da categoria: csv-json → serialization
