+++ src/converters/serialization/pages/serialization-converter-page/serialization-converter-page.component.html
@@ -4,7 +4,7 @@
         <ol>
             <li><a routerLink="/">Início</a></li>
             <li><a routerLink="/conversores">Conversores</a></li>
-            <li><a routerLink="/conversores/csv-json">Serialização de Dados</a></li>
+            <li><a routerLink="/conversores/serialization">Serialização de Dados</a></li>
             <li><span>Converter {{ sourceFormat.name }} para {{ targetFormat.name }}</span></li>
         </ol>
     </nav>
