+++ src/converters/serialization/pages/serialization-home-page/serialization-home-page.component.html
@@ -37,7 +37,7 @@
             <hr>
             <div class="conversion-links-grid">
                 @for (target of group.targets; track target.id) {
-                <a [routerLink]="['/conversores/csv-json', unitUrlFormatterService.generateConversionUrl(group.key.id, target.id)]"
+                <a [routerLink]="['/conversores/serialization', unitUrlFormatterService.generateConversionUrl(group.key.id, target.id)]"
                     class="conversion-link-card">
                     <div class="conversion-link-icon">
                         <span class="material-symbols-outlined">data_object</span>
