+++ src/converters/shared/converters.routes.ts
@@ -19,5 +19,5 @@ export const CONVERTERS_ROUTES: Routes = [
     { path: 'pressao', loadChildren: () => import('../pressure/pressure.routes').then(m => m.PRESSURE_ROUTES) },
     { path: 'velocidade', loadChildren: () => import('../speed/speed.routes').then(m => m.SPEED_ROUTES) },
     { path: 'potencia', loadChildren: () => import('../power/power.routes').then(m => m.POWER_ROUTES) },
-    { path: 'csv-json', loadChildren: () => import('../csv-json/csv-json.routes').then(m => m.CSV_JSON_ROUTES) },
+    { path: 'serialization', loadChildren: () => import('../serialization/serialization.routes').then(m => m.SERIALIZATION_ROUTES) },
 ];
